Extract TabPanel.tsx
