+++ src/App.tsx
@@ -1,8 +1,6 @@
 import * as React from 'react';
-import {useEffect} from 'react';
 import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
-import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
 import {
   Button,
@@ -16,32 +14,7 @@ import {
 import Settings from "./components/Settings";
 import {Info} from "./components/Info";
 import {Assets} from "./components/Assets";
-
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-}
-
-function TabPanel(props: TabPanelProps) {
-  const {children, value, index, ...other} = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box sx={{p: 3}}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  );
-}
+import {TabPanel} from "./components/TabPanel";
 
 function a11yProps(index: number) {
   return {
